+++ src/document/json-document.ts
@@ -96,9 +96,9 @@ export class JSONDocument implements DocumentShared {
                 : value.type === "tuple" || value.type === "array"
                     ? []
                     : value.type === "vector"
-                        ? value.vecType === "FLOAT32"
-                            ? new Float32Array()
-                            : new Float64Array()
+                        ? value.vecType === "FLOAT64"
+                            ? new Float64Array()
+                            : new Float32Array()
                         : undefined);
         }
